+++ fluids/numba.py
@@ -59,7 +59,7 @@ The correct syntax is as follows:
 >>> from fluids.numba import * # May be used without first importing fluids
 '''
 
-caching = False
+caching = True
 __all__ = []
 
 __funcs = {}